Consistent user authentication by IP address, user agent
It is now possible to verify users by IP for trivial things like
downloads, in case the user

	A. Uses an out-of-browser download manager
	B. Streams video using an external video player

Or generally tries to use the website w/o a browser. Of course,
this application should be limited to trivial things like file
downloads.

It's not possible to spoof the REMOTE_ADDR header (I
would be sending the packets to someone else instead of the malicious
recipient!) but still please limit this functionality in case
two users share an IP address or something.